Make everything float32

+++ replay_trajectory_classification/classifier.py
@@ -128,10 +128,13 @@ class _ClassifierBase(BaseEstimator):
                     continuous_state_transition[-1].append(
                         transition.make_state_transition(
                             self.environments, position, is_training,
-                            encoding_group_labels, environment_labels))
+                            encoding_group_labels, environment_labels
+                        ).astype(np.float32)
+                    )
                 else:
                     continuous_state_transition[-1].append(
-                        transition.make_state_transition(self.environments))
+                        transition.make_state_transition(self.environments)
+                        .astype(np.float32))
 
         n_states = len(self.continuous_transition_types)
         self.continuous_state_transition_ = np.zeros(
